requirements.txt
# Data processing 
pandas
numpy

# Web UI (optional)
streamlit

# Development tools
pytest
black
isort
pylint
python-dotenv

# Craw AI
crewai
crewai-tools
duckduckgo-search
langchain-openai 
langchain-community
